Adds automatic dice selection to bets phase
Now, on the bets phase, the best dice are automatically checked to help
the player to identify their combination. This is made via some
functions that we might want to refactor in the future, if only to make
them prettier avoiding the O(n^3) complexity (even though n=5).

I don't know anything about less so the following things remain
pending:
* Ensure that the "Continuar" button appears below the dice.
* Always display the "checked" style for the dice regardless of the
  state, because I think it's useful on all phases.