Adds game tests
For this I brought npm into the project, although it's only needed to
execute the tests.

It currently has a 100% coverage of the game.js file (previously known
as script.js). Although I didn't find any bugs in the actual code, it
made me do some changes to the code structure so now it should be
slightly better.

It would be great to add a GitHub action to ensure that the tests keep
on passing and maybe even to ensure that we don't go below 100%
anymore.

Sorry for the 11K extra lines of package-lock.json, but it's recommended
to keep it in version control.