Fixes initial roll animation
Due to seemingly innocent change, the initial rolling animation broke.
But the real reason is that we're fetching the Mustache template in an
async function so the game state could change between calling the render
function and the actual rendering. I've changed it to ensure that only
the very first time the template is fetched, guaranteeing that
successive calls are done in the same thread.